New Style when sending Emails (#200)
* New Style when sending Emails

After talking with Iara, We came up with this new and more professional Design, when receiving mails from Educado. 
The content "text" "message" has been re-worked, where we added more professional touch, and on top added some more info etc. "If you didn't do this action, please ignore the Mail, or Contact Support"
On top of that, The code is more clean and the Template is now re-usable, Instead of being HardCoded